+++ RELEASE-NOTES.md
@@ -6,6 +6,10 @@
 
 - Updated dependencies [#3317](https://github.com/Automattic/simplenote-electron/pull/3317)
 
+### Fixes
+
+- Fixed paste not working in context menu and enabled it in browsers [#3316](https://github.com/Automattic/simplenote-electron/pull/3316)
+
 ## [v2.23.2]
 
 ### Other Changes